Adds options for config file and log location
This changes makes both the location of the config file and the log
files configurable.

You can set a custom config path at the command line.  By default, it
will be `/etc`.  This value should be a path to a directory.  Within
that directory, the app will look for `aws-signing-proxy.yml` (the file
name itself is still hard coded). The cli option is `--config-location`.

A custom log location can be set at the command line or through the
config file.  By default it will be `/var/log/aws-signing-proxy`.  This
value should also be a path to a directory.  Within the specified
directory, the app will create a `proxy.log` and take care of log
rotation.  The CLI option is `--log-location` and the config entry is
`log-location`.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>